Allowing body to be either string or stream
When this middleware is used together with other middlewares
it may run into a conflict trying to slurp the body.
If the other middleware already slurped the body and put the
resulting string back into the request the slurp in this middleware
fails.
This commit adds handling of string-type body.